requirements.txt
apiclient==1.0.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
pycparser==2.21
PyJWT==1.7.1
pytest==7.4.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.4
